tests/sys/kern/ktls_test: fix -Wsign-compare issue
Cast the size_t quantity used in a comparison to off_t to mute a
`-Wsign-compare` complaint that now occurs after ATF 0.22 [1].

MFC after:	2 weeks
Reported by:	clang
Reviewed by:	markj
Differential Revision:	https://reviews.freebsd.org/D59285
[1]: https://github.com/freebsd/atf/pull/72

+++ tests/sys/kern/ktls_test.c
@@ -3011,7 +3011,7 @@ ATF_TC_BODY(ktls_receive_loopback_sendfile, tc)
 	debug_hexdump(tc, p, payload_len, "shm buffer before");
 	ATF_REQUIRE_INTEQ(0, sendfile(shm, sockets[1], 0, payload_len, &hdtr,
 	    &sbytes, 0));
-	ATF_REQUIRE_INTEQ(sbytes, len);
+	ATF_REQUIRE_INTEQ(sbytes, (off_t)len);
 
 	outbuf = calloc(payload_len, 1);
